use history API instead of setting location hash
Fixes #554 - page now scrolls smoothly to the tab bar when a tab is
clicked in Firefox, and there's no longer that strange little jump when
clicking a tab a second time.

Remove save/restore scroll workaround and use the History API instead
of setting the location hash (unless API is not supported).