remove Firefox sniffing in smoothScroll
Use `window.scrollTo()` and `window.[scrollY|pageYOffset]` instead of `[documentElement|body].scrollTop` to achieve cross-browser scrolling without the need to sniff Firefox.

Also, simplify the inaccurracy fix at the end of the scroll by scrolling directly to the target position (known from the start) instead of looking for the remaining offset by which to scroll the page.